+++ zyte_spider_templates/__init__.py
@@ -1,2 +1,3 @@
-from .spiders.base import BaseSpider, BaseSpiderParams
+from .params import make_params
+from .spiders.base import BaseSpider
 from .spiders.ecommerce import EcommerceSpider
