+++ zyte_spider_templates/spiders/base.py
@@ -1,53 +1,11 @@
 from importlib.metadata import version
-from typing import Any, Dict, Optional
+from typing import Any, Dict
 
 import scrapy
-from pydantic import BaseModel, Field
 from scrapy.crawler import Crawler
-
-from zyte_spider_templates._geolocations import (
-    GEOLOCATION_OPTIONS_WITH_CODE,
-    Geolocation,
-)
 
 # Higher priority than command-line-defined settings (40).
 ARG_SETTING_PRIORITY: int = 50
-
-
-class BaseSpiderParams(BaseModel):
-    url: str = Field(
-        title="URL",
-        description="Initial URL for the crawl. Enter the full URL including http(s), "
-        "you can copy and paste it from your browser. Example: https://toscrape.com/",
-        pattern=r"^https?://[^:/\s]+(:\d{1,5})?(/[^\s]*)*(#[^\s]*)?$",
-    )
-    geolocation: Optional[Geolocation] = Field(
-        title="Geolocation",
-        description="ISO 3166-1 alpha-2 2-character string specified in "
-        "https://docs.zyte.com/zyte-api/usage/reference.html#operation/extract/request/geolocation.",
-        default=None,
-        json_schema_extra={
-            "enumMeta": {
-                code: {
-                    "title": GEOLOCATION_OPTIONS_WITH_CODE[code],
-                }
-                for code in Geolocation
-            }
-        },
-    )
-    max_requests: Optional[int] = Field(
-        description=(
-            "The maximum number of Zyte API requests allowed for the crawl.\n"
-            "\n"
-            "Requests with error responses that cannot be retried or exceed "
-            "their retry limit also count here, but they incur in no costs "
-            "and do not increase the request count in Scrapy Cloud."
-        ),
-        default=100,
-        json_schema_extra={
-            "widget": "request-limit",
-        },
-    )
 
 
 class BaseSpider(scrapy.Spider):
